% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_text.R
\name{flow_text_from_file}
\alias{flow_text_from_file}
\title{wrapper for batch reading data from file}
\usage{
flow_text_from_file(generator = text_data_generator())
}
\description{
wrapper for batch reading data from file
}
