% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_text.R
\name{flow_text_to_file}
\alias{flow_text_to_file}
\title{flow text from a textdf to an input and output file}
\usage{
flow_text_to_file(textdf, nchars = 20, outputs, filepath = tempdir(),
  append = FALSE, filenames = c("inputs.txt", "outputs.txt"))
}
\description{
flow text from a textdf to an input and output file
}
