% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_text.R
\name{text_data_generator}
\alias{text_data_generator}
\title{generate a textflow.batchFlow connection from an input and output file into a model
generate a textflow.batchFlow connection from an input and output file into a model}
\usage{
text_data_generator(directory = ".", inputfile = "inputs.txt",
  outputfile = "outputs.txt", nchars = 20, batchSize = 32,
  inputs = "")
}
\description{
generate a textflow.batchFlow connection from an input and output file into a model
generate a textflow.batchFlow connection from an input and output file into a model
}
