% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textflow.batchFlow.R
\docType{data}
\name{textflow.batchFlow}
\alias{textflow.batchFlow}
\title{wrapper for batch reading data from file}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
textflow.batchFlow
}
\description{
wrapper for batch reading data from file
}
\keyword{datasets}
